requirements.txt
streamlit
openai
llama-index
nltk
transformers
torch
python-dotenv